# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Daily Challenge Bingo API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  participant: 'participant',
} as const;

export interface User {
  id: number;
  name: string;
  username: string;
  /** @nullable */
  email?: string | null;
  role: UserRole;
}

export type ParticipantRole = typeof ParticipantRole[keyof typeof ParticipantRole];


export const ParticipantRole = {
  admin: 'admin',
  participant: 'participant',
} as const;

export interface Participant {
  id: number;
  name: string;
  username: string;
  /** @nullable */
  email?: string | null;
  role: ParticipantRole;
  createdAt: string;
}

export type ParticipantInputRole = typeof ParticipantInputRole[keyof typeof ParticipantInputRole];


export const ParticipantInputRole = {
  admin: 'admin',
  participant: 'participant',
} as const;

export interface ParticipantInput {
  name: string;
  username: string;
  password: string;
  email?: string;
  role?: ParticipantInputRole;
}

export interface ParticipantUpdate {
  name?: string;
  username?: string;
  password?: string;
  email?: string;
}

export interface ParticipantProgress {
  participantId: number;
  name: string;
  totalBoxes: number;
  hidden: number;
  revealed: number;
  completed: number;
  reflectionsCount: number;
}

export interface BingoCard {
  id: number;
  userId: number;
  participantName?: string;
  title: string;
  rows: number;
  cols: number;
  createdAt: string;
}

export type BingoBoxInputCategory = typeof BingoBoxInputCategory[keyof typeof BingoBoxInputCategory];


export const BingoBoxInputCategory = {
  Spirit: 'Spirit',
  Mind: 'Mind',
  Body: 'Body',
  Health: 'Health',
  People: 'People',
} as const;

export interface BingoBoxInput {
  boxNumber: number;
  category: BingoBoxInputCategory;
  challengeText: string;
}

export interface BingoCardInput {
  userId: number;
  title: string;
  boxes: BingoBoxInput[];
}

export interface BingoCardUpdate {
  title?: string;
}

export type BingoBoxCategory = typeof BingoBoxCategory[keyof typeof BingoBoxCategory];


export const BingoBoxCategory = {
  Spirit: 'Spirit',
  Mind: 'Mind',
  Body: 'Body',
  Health: 'Health',
  People: 'People',
} as const;

export interface BingoBox {
  id: number;
  cardId: number;
  boxNumber: number;
  category: BingoBoxCategory;
  challengeText: string;
  isRevealed: boolean;
  isCompleted: boolean;
  /** @nullable */
  revealedAt?: string | null;
  /** @nullable */
  completedAt?: string | null;
}

export interface BingoCardWithBoxes {
  id: number;
  userId: number;
  participantName?: string;
  title: string;
  rows: number;
  cols: number;
  createdAt: string;
  boxes: BingoBox[];
}

export type BingoBoxUpdateCategory = typeof BingoBoxUpdateCategory[keyof typeof BingoBoxUpdateCategory];


export const BingoBoxUpdateCategory = {
  Spirit: 'Spirit',
  Mind: 'Mind',
  Body: 'Body',
  Health: 'Health',
  People: 'People',
} as const;

export interface BingoBoxUpdate {
  category?: BingoBoxUpdateCategory;
  challengeText?: string;
  isRevealed?: boolean;
  isCompleted?: boolean;
}

export interface Reflection {
  id: number;
  userId: number;
  participantName?: string;
  date: string;
  whatIChose: string;
  whatIDid: string;
  impact: string;
  createdAt: string;
}

export interface ReflectionInput {
  whatIChose: string;
  whatIDid: string;
  impact: string;
}

export interface ReflectionUpdate {
  whatIChose?: string;
  whatIDid?: string;
  impact?: string;
}

export type DashboardSummaryRole = typeof DashboardSummaryRole[keyof typeof DashboardSummaryRole];


export const DashboardSummaryRole = {
  admin: 'admin',
  participant: 'participant',
} as const;

export interface DashboardSummary {
  role: DashboardSummaryRole;
  /** @nullable */
  totalParticipants?: number | null;
  /** @nullable */
  totalCards?: number | null;
  /** @nullable */
  totalBoxesRevealed?: number | null;
  /** @nullable */
  totalBoxesCompleted?: number | null;
  /** @nullable */
  totalReflections?: number | null;
  recentReflections?: Reflection[];
  /** @nullable */
  myBoxesTotal?: number | null;
  /** @nullable */
  myBoxesRevealed?: number | null;
  /** @nullable */
  myBoxesCompleted?: number | null;
  /** @nullable */
  myReflectionsCount?: number | null;
  /** @nullable */
  hasReflectedToday?: boolean | null;
}