# lib/api-zod/src/generated/types/participantInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Daily Challenge Bingo API
 * OpenAPI spec version: 0.1.0
 */
import type { ParticipantInputRole } from './participantInputRole';

export interface ParticipantInput {
  name: string;
  username: string;
  password: string;
  email?: string;
  role?: ParticipantInputRole;
}